// src - SiteFrontend avc ResptAPI-SpringBoot/app/models/product.model.ts
export class ProductModel {
  id?: any ;
  name?: any  ;
  description?: string ;
  price?: any ;
  quantity?: number ;
  colorsAvailaibles?: string  ;
  sizesAvailaibles?: string ;
  addDate?: Date ;
  categoryName?: string  ;
  populary?: boolean ;
  newProduct?: boolean ;
  particularity?: string ;
  image?: any ;
  image2?: any ;
  image3?: any ;
  image4?: any ;
  image5?: any ;

}